refactor(xp-tracker): refactored xp parser and diff to not care about specific ranges
