+++ src/app/services/hiscores/hiscore.enum.ts
@@ -1,0 +1,109 @@
+export enum SkillEnum {
+  Overall = 'Overall',
+  Attack = 'Attack',
+  Defence = 'Defence',
+  Strength = 'Strength',
+  Hitpoints = 'Hitpoints',
+  Ranged = 'Ranged',
+  Prayer = 'Prayer',
+  Magic = 'Magic',
+  Cooking = 'Cooking',
+  Woodcutting = 'Woodcutting',
+  Fletching = 'Fletching',
+  Fishing = 'Fishing',
+  Firemaking = 'Firemaking',
+  Crafting = 'Crafting',
+  Smithing = 'Smithing',
+  Mining = 'Mining',
+  Herblore = 'Herblore',
+  Agility = 'Agility',
+  Thieving = 'Thieving',
+  Slayer = 'Slayer',
+  Farming = 'Farming',
+  Runecraft = 'Runecraft',
+  Hunter = 'Hunter',
+  Construction = 'Construction',
+}
+
+export enum CompetitiveEnum {
+  LeaguePoints = 'League Points',
+  PvpArena = 'PvP Arena - Rank',
+  LastManStanding = 'Last Man Standing (Rank)',
+}
+
+export enum BountyHunterEnum {
+  BountyHunter = 'Bounty Hunter - Hunter',
+  BountyHunterRogues = 'Bounty Hunter - Rogues',
+}
+
+export enum ClueScrollsEnum {
+  ClueScrollsAll = 'Clue Scrolls (all)',
+  ClueScrollsBeginner = 'Clue Scrolls (beginner)',
+  ClueScrollsEasy = 'Clue Scrolls (easy)',
+  ClueScrollsMedium = 'Clue Scrolls (medium)',
+  ClueScrollsHard = 'Clue Scrolls (hard)',
+  ClueScrollsElite = 'Clue Scrolls (elite)',
+  ClueScrollsMaster = 'Clue Scrolls (master)',
+}
+
+export enum MiniGameEnum {
+  SoulWarsZeal = 'Soul Wars Zeal',
+  RiftsClosed = 'Rifts Closed',
+}
+
+export enum RaidEnum {
+  ChambersOfXeric = 'Chambers of Xeric',
+  ChambersOfXericChallengeMode = 'Chambers of Xeric: Challenge Mode',
+  TheGauntlet = 'TheGauntlet',
+  TheCorruptedGauntlet = 'The Corrupted Gauntlet',
+  TheathreOfBlood = 'Theatre of Blood',
+  TheathreOfBloodHardMode = 'Theatre of Blood: Hard Mode',
+  TombsOfAmascut = 'Tombs of Amascut',
+  TombsOfAmascutExpertMode = 'Tombs of Amascut: Expert Mode',
+}
+
+export enum BossEnum {
+  AbyssalSire = 'Abyssal Sire',
+  AlchemicalHydra = 'Alchemical Hydra',
+  BarrowsChests = 'Barrows Chests',
+  Bryophyta = 'Bryophyta',
+  Callisto = 'Callisto',
+  Cerberus = 'Cerberus',
+  ChaosElemental = 'Chaos Elemental',
+  ChaosFanatic = 'Chaos Fanatic',
+  CommanderZilyana = 'Commander Zilyana',
+  CorporealBeast = 'Corporeal Beast',
+  CrazyArchaeologist = 'Crazy Archaeologist',
+  DagannothPrime = 'Dagannoth Prime',
+  DagannothRex = 'Dagannoth Rex',
+  DagannothSupreme = 'Dagannoth Supreme',
+  DerangedArchaeologist = 'Deranged Archaeologist',
+  GeneralGraardor = 'General Graardor',
+  GiantMole = 'Giant Mole',
+  GrotesqueGuardians = 'Grotesque Guardians',
+  Hespori = 'Hespori',
+  KalphiteQueen = 'Kalphite Queen',
+  KingBlackDragon = 'King Black Dragon',
+  Kraken = 'Kraken',
+  KreeAra = "Kree'Arra",
+  KrilTsutsaroth = "K'ril Tsutsaroth",
+  Mimic = 'Mimic',
+  Nex = 'Nex',
+  Nightmare = 'Nightmare',
+  PhosanisNightmare = "Phosani's Nightmare",
+  Obor = 'Obor',
+  PhantomMuspah = 'Phantom Muspah',
+  Sarachnis = 'Sarachnis',
+  Scorpia = 'Scorpia',
+  Skotizo = 'Skotizo',
+  Tempoross = 'Tempoross',
+  ThermoNuclearSmokeDevil = 'Thermonuclear Smoke Devil',
+  TzKalZuk = 'TzKal-Zuk',
+  TzTokJad = 'TzTok-Jad',
+  Venenatis = 'Venenatis',
+  Vetion = "Vet'ion",
+  Vorkath = 'Vorkath',
+  Wintertodt = 'Wintertodt',
+  Zalcano = 'Zalcano',
+  Zulrah = 'Zulrah',
+}
